Assignment2
With the successful update of 300 Windows servers to version 2022. Your supervisor is asking you to come up with a plan on a semi-automated patch delivery system. Your organization has a lot of software that requires maintenance through patches, and these patches are important for ensuring security and stability of all the applications and hardware within your infrastructure. 
You task is to come up with a plan for rolling out patches? 
For this assignment you will breakdown the components that require patching and provide schedule on how these patches will be deployed and implemented. 
Use the following questions to help guide how you start thinking about the various patches that you have to cover:  
•	Operating System patches (Environment has Windows and Linux Operating Systems)
•	Application patches (Various software application such as from Microsoft products, Linux web-servers etc..) 
•	Hardware patches (routers/switches/access points etc..) 
•	End-user workstation patches (Windows 10/11 Operating Systems, Ubuntu Linux Desktop environments etc..) 
Part of this plan is also to ensure that respective notifications are sent to the teams/users that will have impacts of these patches. Please ensure that you incorporate what methods of notifications will be included.