Assignment3
As IT environments evolve, it is important to understand the need to implement changes to the ensure continued uptime whilst also providing improvements, bug fixes to the user base. This is where the change management process comes into place. 
Your direct supervisor has come to you to get a production software change done to the production Microsoft SQL database server, this change involves updating a stored procedure and adding an index to the respective tables. 
This request has come done from key stakeholders in the business as it improves the speed at the data is queried and returned. 
Additional Details: 
This Microsoft SQL server contains multiple databases that various applications that is used by the business.  
This change will only affect one database, however, there is a risk of potential downtime and access restriction to the MS SQL server itself. 
You task is to come up with a plan to get this change scheduled into the production environment.
Key items to consider in your plan:
1) Will you need to get approvals for this change? What mechanism will you use? 
2) Is there a need to have this change tested? If so, who will be responsible for this?
3) How will you validate that the change will not impact the other databases or the applications that interact with those databases?
4) Will an outage need to be declared, if so, how big of a window will be needed?
5) What will be your success criteria?
6) What other areas/items should you consider?