Assignment4
This assignment builds upon Assignment 3. 
In the previous assignment you were introduced to the change management process for the Microsoft SQL database, the entire process led to additional questions being asked about reducing downtime for future changes or perhaps another way to access the data during these changes. 
Your supervisor comes to you and asks you to investigate and plan out using a cloud-based database service. For simplicity, your environment is using Microsoft Azure as it’s cloud platform. 
Your task is to come up with a plan on how you are going to plan this type of integration. 
Key Items to include in your plan: 
1)	Which method do you prefer to deploy? A cloud hosted MS SQL server or using the Azure SQL service? Please explain your decision and provide supporting material in your decision.

2)	How will you plan out data -synchronization between the on-premise databases to the cloud-bases databases? 

3)	What will be your process in the event to have the business users point to the cloud-based solution when the on-premises environment is down due to updates, or due to an outage?